WIP exporting working
