+++ lib/cgate-database.js
@@ -7,11 +7,11 @@ require('../hot-debug.js');
 const log = require('debug')('cbus:db');
 
 const _ = require('lodash');
+const chalk = require('chalk');
 const bytes = require('bytes');
 
 const xml2js = require('xml2js');
 const lodash = require('lodash');
-const stringify = require('json-stable-stringify');
 
 const cbusUtils = require('./cbus-utils.js');
 const CBusNetId = require('./cbus-netid.js');
@@ -19,6 +19,8 @@ const CBusNetId = require('./cbus-netid.js');
 module.exports = CGateDatabase;
 
 let UNIT_TYPES;
+
+const DEFAULT_CBUS_APPLICATION = 56;
 
 // CGateDatabase loads the objeect database from C-Gate, parsing it into three lists:
 // * applications
@@ -29,241 +31,16 @@ function CGateDatabase(netId) {
 	console.assert(netId.isProjectId());
 	this.netId = netId;
 
-	// array of applications
+	// maps
 	this.applications = undefined;
-
-	// array of groups
 	this.groups = undefined;
-
-	// array of units
 	this.units = undefined;
+	this.unrecognisedUnits = undefined;
 
 	// unit-type lookup
 	const typesJSON = fs.readFileSync(`./resources/unit-types.json`);
 	UNIT_TYPES = JSON.parse(typesJSON);
-	console.log(util.inspect(UNIT_TYPES, { breakLength: 100, colors:true}));
 }
-
-function _getScaffolding(platform) {
-	const config = _.clone(platform.config);
-	config.accessories = [`ACCESSORIES_PLACEHOLDER`];
-
-	let data = {
-		platforms: [config]
-	};
-
-	let compare = function (keyA, keyB) {
-		// determine the key order
-		function getRank(key) {
-			const KEY_RANK = [
-				`platform`, `name`,
-				`client_ip_address`, `client_controlport`,
-				`client_cbusname`, `client_network`, `client_application`,
-				`platform_export`,
-				`application`
-			];
-			let rank = _.findIndex(KEY_RANK, o => o === key);
-			return (rank === -1) ? KEY_RANK.length : rank;
-		}
-
-		return getRank(keyA.key) > getRank(keyB.key);
-	};
-
-	return stringify(data, { cmp: compare, space: 2 })
-		.replace(/^      "accessories": /mg, `\n      "accessories": `)
-		.replace(/^      "platform_export": /mg, `\n      "platform_export": `);
-}
-
-function _flattenAccessories(accessories) {
-	let compare = function (keyA, keyB) {
-		// determine the key order, or KEY_ORDER.length if not found
-		function getRank(key) {
-			const KEY_RANK = [`netId`, `type`, `network`, `application`, `id`, `name`, `dbTag`, `invert`, `activeDuration`, `enabled`];
-			const rank = _.findIndex(KEY_RANK, o => o === key);
-			return (rank === -1) ? KEY_RANK.length : rank;
-		}
-
-		return getRank(keyA.key) > getRank(keyB.key);
-	};
-
-	let flattened = stringify(accessories, {cmp: compare});
-
-	// remove the square backets around the entire string
-	flattened = flattened.slice(1, flattened.length - 1);
-
-	if (flattened.length !== 0) {
-		flattened = flattened.replace(/\},/g, `},\n`)		// put a newline after all `},`
-			.replace(/\{/g, `{ `)			// put a space after all `{`
-			.replace(/\}/g, ` }`)			// put a space before all `}`
-			.replace(/":/g, `": `)			// put space after colons
-			.replace(/,"/g, `, "`)			// put space between commas and quotes (property key start)
-			.replace(/^/gm, `        `);	// indent lines
-	}
-
-	return flattened;
-}
-
-/**
- * merge the database with the accessory list
- */
-CGateDatabase.prototype._buildUnifiedAccessoryList = function (platform) {
-	let accessories = [];
-
-	// create accessory definitions from db
-	_.forOwn(this.groups, (group, netIdStr) => {
-		// build up the definition of the platform (minus the accessories
-		let dbNetId = CBusNetId.parse(netIdStr);
-		let accessory = {
-			type: `unknown`,
-			network: dbNetId.network,
-			application: dbNetId.application,
-			id: dbNetId.group,
-			name: group.tag,
-			enabled: false,
-			_netId: dbNetId,
-		};
-
-		accessories.push(accessory);
-	});
-
-	// iterate through loaded accessories, adding/modifying entries as necessary
-	_.forEach(platform.config.accessories, config => {
-		const configNetId = new CBusNetId(
-			platform.config.client_cbusname,
-			config.network || platform.config.client_network,
-			config.application || platform.config.client_application,
-			config.id);
-
-		let found = _.find(accessories, o => o._netId.isEquals(configNetId));
-
-		// create new shadow
-		if (found) {
-			// modify found to match
-			if (found.name !== config.name) {
-				// stash away dbTag
-				found.dbtag = found.name;
-				found.name = config.name;
-			}
-			found.type = config.type;
-			found.enabled = (config.enabled !== false);
-		} else {
-			let shadow = _.clone(config);
-			shadow._netId = configNetId;
-			accessories.push(shadow);
-		}
-	});
-
-	// removing network and application properties where unnecessary
-	const defaultNetId = new CBusNetId(this.netId.project, platform.config.client_network, platform.config.client_application);
-	_.forEach(accessories, shadow => {
-		if (shadow._netId.isSameNetwork(defaultNetId)) {
-			delete shadow.network;
-		}
-
-		if (shadow._netId.isSameApplication(defaultNetId)) {
-			delete shadow.application;
-		}
-
-		if (shadow.enabled === true) {
-			delete shadow.enabled;
-		}
-
-		shadow._hash = shadow._netId.getHash();
-	});
-
-	// sort by _hash
-	accessories = _.sortBy(accessories, o => o._hash);
-
-	// remove _hash property
-	_.forEach(accessories, o => {
-		delete o._hash;
-		delete o._netId;
-	});
-
-	return accessories;
-};
-
-CGateDatabase.prototype.stringifyPlatform = function (platform) {
-	let scaffolding = _getScaffolding(platform);
-
-	// build accessory section
-	let accessories = this._buildUnifiedAccessoryList(platform);
-
-	const [enabled, disabled] = _.partition(accessories, o => (o.enabled !== false));
-	const flatEnabled = _flattenAccessories(enabled);
-	const flatDisabled = _flattenAccessories(disabled);
-
-	let accStr = '';
-	if (flatEnabled.length > 0) {
-		accStr += flatEnabled;
-		if (flatDisabled.length > 0) {
-			accStr += `,\n\n`;
-		}
-	}
-
-	if (flatDisabled.length > 0) {
-		accStr += flatDisabled;
-	}
-
-	// merge accessories into scaffolding
-	return scaffolding.replace(/^ +"ACCESSORIES_PLACEHOLDER"$/gm, accStr) + `\n`;
-};
-
-CGateDatabase.prototype.exportPlatform = function (path, platform, callback) {
-	const output = this.stringifyPlatform(platform);
-
-	fs.writeFile(path, output, function (err) {
-		if (err) {
-			log(`Platform failed to export: ${err}`);
-		} else {
-			const numBytes = fs.statSync(path).size;
-			const byteStr = bytes(numBytes, {unitSeparator:' '});
-			log(`Platform exported to ${path} (${byteStr})`);
-
-			if (callback) {
-				callback(err);
-			}
-		}
-	});
-};
-
-CGateDatabase.prototype.stringifyDatabase = function () {
-	const data = {
-		applications: this.applications,
-		groups: this.groups,
-		units: this.units
-	};
-
-	let compare = function (a, b) {
-		if (a.key.startsWith(`//`) && b.key.startsWith(`//`)) {
-			const aId = CBusNetId.parse(a.key);
-			const bId = CBusNetId.parse(b.key);
-
-			return CBusNetId.compare(aId, bId);
-		}
-
-		return a.key < b.key ? -1 : 1;
-	};
-
-	return stringify(data, {cmp: compare, space: 2});
-};
-
-CGateDatabase.prototype.exportDatabase = function (path, callback) {
-	const jsonStr = this.stringifyDatabase();
-
-	fs.writeFile(path, jsonStr, function (err) {
-		if (err) {
-			log(`Database failed to export: ${err}`);
-		} else {
-			const numBytes = fs.statSync(path).size;
-			log(`Database exported to ${path} (${bytes(numBytes, {unitSeparator:' '})})`);
-		}
-
-		if (callback) {
-			callback(err);
-		}
-	});
-};
 
 CGateDatabase.prototype.getStats = function () {
 	return {
@@ -289,12 +66,18 @@ CGateDatabase.prototype.fetch = function (client, callback) {
 			console.assert(!err, `dbgetxml parse failure`, err);
 			const result = _parseXML(databaseXML, new CBusNetId(this.netId.project, this.netId.network));
 
-			// patch up types
-			_patchDevicesToGroups(result);
-
 			this.applications = result.applications;
 			this.groups = result.groups;
 			this.units = result.units;
+			this.unrecognisedUnits = result.unrecognisedUnits;
+
+			if (!_.isEmpty(result.unrecognisedUnits)) {
+				log(`${chalk.red(`WARNING`)}: unknown unit types encountered: `);
+				log(util.inspect(result.unrecognisedUnits, {colors: true}));
+			}
+
+			// patch up types
+			this._patchDevicesToGroups(result);
 
 			if (callback) {
 				callback();
@@ -303,28 +86,7 @@ CGateDatabase.prototype.fetch = function (client, callback) {
 	});
 };
 
-/**
- * create a result entity that contains 3 arrays: applications, groups, units
- */
-function _parseXML(databaseXML) {
-	let types = {};
-
-	const result = {
-		applications: {},
-		groups: {},
-		units: {}
-	};
-
-	const project = databaseXML.installation.project;
-	const projectName = project.address;
-
-	_.forEach(project.network, network => _parseNetwork(network, projectName, result));
-
-	return result;
-}
-
-function _parseNetwork(network, projectName, result) {
-	// build object for applications and groups
+function _extractApplicationsAndGroups (network, projectName, result) {
 	_arrayize(network.application).forEach(srcApp => {
 		const appId = new CBusNetId(projectName, network.address, cbusUtils.integerise(srcApp.address));
 		result.applications[appId.toString()] = {
@@ -347,8 +109,9 @@ function _parseNetwork(network, projectName, result) {
 			});
 		}
 	});
+}
 
-	// build object for physical units
+function _extractUnits(network, projectName, result) {
 	_arrayize(network.unit).forEach(srcUnit => {
 		// find the property section listing wanted attributes
 		// in: <PP Name="GroupAddress" Value="0x19 0x1a 0x1b 0x1d 0x33 0x34 0x1f 0x1e 0x21 0x22 0x23 0x24 0xff 0xff 0xff 0xff"/>
@@ -356,9 +119,9 @@ function _parseNetwork(network, projectName, result) {
 		function extractProperty(propertyName) {
 			let result = srcUnit.properties[propertyName]; // `GroupAddress`;
 
-			if (typeof result === `undefined`) {
-				log(`WARNING: parameter ${propertyName} missing in ${util.inspect(srcUnit)}`);
-			}
+			// if (typeof result === `undefined`) {
+			// 	log(`WARNING: parameter ${propertyName} missing in ${util.inspect(srcUnit)}`);
+			// }
 
 			return result;
 		}
@@ -378,13 +141,17 @@ function _parseNetwork(network, projectName, result) {
 			return result;
 		}
 
-		function _extractGroupAddresses(unit) {
+		function _extractGroupAddresses(unit, unitId) {
 			let groups;
 
 			// only keep as many groups as we expect
 			const unitInfo = UNIT_TYPES[srcUnit.unittype];
 			if (typeof unitInfo === `undefined`) {
-				log(`WARNING: unknown unit type: ${srcUnit.unittype} -- add to resources/unit-types.json`);
+				// record unrecognised unittype
+				if (typeof result.unrecognisedUnits[srcUnit.unittype] === `undefined`) {
+					result.unrecognisedUnits[srcUnit.unittype] = [];
+				}
+				result.unrecognisedUnits[srcUnit.unittype].push({tag: unit.tag, netId: unitId});
 			} else {
 				let numGroups = unitInfo.outputChannels;
 				if (typeof numGroups !== `undefined`) {
@@ -392,14 +159,15 @@ function _parseNetwork(network, projectName, result) {
 					groups = _.slice(groups, 0, numGroups);
 				}
 
+				// it's possible that the application has not been set, if so, default to the lighting application
+				// TODO check that this is the right behaviour
 				if (typeof unit.application !== `number`) {
-					unit.application = 57;	// default
-					// log(`fuck! ` + util.inspect(unit, { breakLength: 100, colors:true}));
+					unit.application = DEFAULT_CBUS_APPLICATION;
 				}
 				console.assert(typeof unit.application === `number`);
 
-				// replace address with netId ('255' with null)
-				groups = _.map(groups, o => (o === 255) ? null : new CBusNetId(projectName, network.address, unit.application, cbusUtils.integerise(o)));
+				// replace address with string representation of netId ('255' with null)
+				groups = _.map(groups, o => (o === 255) ? null : new CBusNetId(projectName, network.address, unit.application, cbusUtils.integerise(o)).toString());
 			}
 
 			return groups;
@@ -424,13 +192,11 @@ function _parseNetwork(network, projectName, result) {
 			serialNumber: srcUnit.serialnumber,
 			catalogNumber: srcUnit.catalognumber,
 			unitType: srcUnit.unittype,
-			// project: extractString(`Project`).trim(),
-			// network: _.nth(extractNumbers(`NetworkAddress`), 0),
-			application: _.nth(extractNumbers(`Application`),0)
+			application: _.nth(extractNumbers(`Application`), 0)
 		};
 
 		// process group addresses
-		unit.groups = _extractGroupAddresses(unit);
+		unit.groups = _extractGroupAddresses(unit, netId.toString());
 
 		// asssign the inferred type
 		const unitInfo = UNIT_TYPES[srcUnit.unittype];
@@ -438,32 +204,53 @@ function _parseNetwork(network, projectName, result) {
 			unit.inferredType = unitInfo.type;
 		}
 
-		log(util.inspect(unit));
-
-		// if (typeof types[unit.unitType] === `undefined`) {
-		// 	types[unit.unitType] = [];
-		// }
-		// types[unit.unitType].push(unit.groups);
-
 		result.units[netId.toString()] = unit;
 	});
-
-	// log(util.inspect(types, { breakLength: 100, colors:true}));
 }
 
-function _patchDevicesToGroups(tree) {
-	//tree.groups;
-	_.forEach(tree.units, (unit, unitId) => {
-		log(`\nprocessing unit ${unitId} (${unit.unitType} -> ${unit.inferredType})`);
-		log(util.inspect(unit, { breakLength: 100, colors:true}));
+/**
+ * create a result entity that contains the arrays:
+ * applications:
+ * groups:
+ * units:
+ * unrecognisedUnits: object mapping type: [ unit]
+ */
+function _parseXML(databaseXML) {
+	const result = {
+		applications: {},
+		groups: {},
+		units: {},
+		unrecognisedUnits: {}
+	};
 
-		if (Array.isArray(unit.groups)) {
-			_.forEach(unit.groups, (groupId, channel) => {
-				log(`linking unit ${unitId}/${channel} to group ${groupId}`);
+	const project = databaseXML.installation.project;
+	_.forEach(project.network, network => {
+		_extractApplicationsAndGroups(network, project.address, result);
+		_extractUnits(network, project.address, result);
+	});
+
+	return result;
+}
+
+CGateDatabase.prototype._patchDevicesToGroups = function () {
+	function link(unit, unitId, channelNumber, group, groupId, type) {
+		log(`linking ${chalk.red.bold(unit.tag)}/${channelNumber} (${unitId}) as ${type} to ${chalk.red.bold(group ? group.tag : `unknown`)} (${groupId})`);
+	}
+
+	_.forEach(this.units, (unit, unitId) => {
+		if (!_.isEmpty(_.compact(unit.groups))) { // (Array.isArray(unit.groups)) {
+			log(`\nprocessing unit ${unitId} (${unit.unitType} -> ${unit.inferredType})`);
+			log(util.inspect(unit, { breakLength: 100, colors:true}));
+
+			_.forEach(unit.groups, (groupId, channelInd) => {
+				if (groupId) {
+					const group = this.getNetworkEntity(groupId);
+					link(unit, unitId, channelInd + 1, group, groupId, unit.inferredType);
+				}
 			});
 		}
 	});
-}
+};
 
 function _arrayize(element) {
 	return Array.isArray(element) ? element : [element];
@@ -486,11 +273,13 @@ CGateDatabase.prototype.getNetworkEntity = function (netId) {
 	if (typeof this.applications === `undefined`) {
 		return undefined;
 	}
-	console.assert(this.groupMap !== `undefined`, `if we have this.applications, then we should have this.groupMap`);
 
-	// TODO perhaps change to throw?
+	if (typeof netId === `string`) {
+		netId = CBusNetId.parse(netId);
+	}
+
 	if (this.netId.project !== netId.project) {
-		throw new Error(`getGroupName unable to search outside default project`);
+		throw new Error(`getNetworkEntity unable to search outside default project`);
 	}
 
 	let result;
