+++ package.json
@@ -30,12 +30,13 @@
   },
   "dependencies": {
     "bytes": "^2.4.0",
-    "ms": "^0.7.2",
     "carrier": "^0.3.0",
     "chalk": "^1.1.3",
     "debug": ">=2.4.0",
     "json-stable-stringify": "^1.0.1",
     "lodash": "^4.17.4",
+    "moment": "^2.18.1",
+    "ms": "^0.7.2",
     "xml2js": "^0.4.17"
   },
   "devDependencies": {
